requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
openai==2.4.0
httpx==0.28.1
pydantic==2.12.2
python-dotenv==1.1.1
requests==2.32.5